+++ Sources/WXKDarkSky/DataBlock.swift
@@ -1,0 +1,8 @@
+//
+//  DataBlock.swift
+//  WXKDarkSkyPackageDescription
+//
+//  Created by Jonathan Thornton on 1/3/18.
+//
+
+import Foundation
