+++ Sources/WXKDarkSky/Language.swift
@@ -1,0 +1,8 @@
+//
+//  Language.swift
+//  WXKDarkSkyPackageDescription
+//
+//  Created by Jonathan Thornton on 1/3/18.
+//
+
+import Foundation
