+++ Sources/WXKDarkSky/Point.swift
@@ -1,0 +1,8 @@
+//
+//  Point.swift
+//  WXKDarkSkyPackageDescription
+//
+//  Created by Jonathan Thornton on 1/3/18.
+//
+
+import Foundation
